% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_psm.R
\name{build_msstats}
\alias{build_msstats}
\title{Conversion of screened peptide quantification files to MSstatsTMT objects for estimating protein intensities. This function only works with tmt data.}
\usage{
build_msstats(quant, annotation)
}
\arguments{
\item{quant}{Filtered peptide intensity, filter_psm output result}

\item{annotation}{experiment_annotation.tsv file from fragpipe}
}
\value{
A list contains two objects, list[[1]] is an MSstatsTMT object that can be used for quality control and subsequent analyses, list[[2]] is the estimated protein intensity obtained using the MSstatsTMT algorithm
}
\description{
Conversion of screened peptide quantification files to MSstatsTMT objects for estimating protein intensities. This function only works with tmt data.
}
\examples{
This function only works with tmt data.


}
