% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_psm.R
\name{build_msstats_lbf}
\alias{build_msstats_lbf}
\title{Conversion of screened peptide quantification files to MSstats objects for estimating protein intensities. This function only works with label-free data.}
\usage{
build_msstats_lbf(quant, ...)
}
\arguments{
\item{quant}{Filtered peptide intensity, filter_psm output result}

\item{...}{}
}
\value{
A list contains two objects, list[[1]] is an MSstats object that can be used for quality control and subsequent analyses, list[[2]] is the estimated protein intensity obtained using the MSstats algorithm
}
\description{
Conversion of screened peptide quantification files to MSstats objects for estimating protein intensities. This function only works with label-free data.
}
\seealso{
\code{\link{MSstats::dataProcess}}
}
