% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_psm.R
\name{dia_quant}
\alias{dia_quant}
\title{Use this function to merge data from different fractions}
\usage{
dia_quant(quant, meta, merge)
}
\arguments{
\item{quant}{filter_dia_psm output}

\item{meta}{fp-manifest file}

\item{merge}{T or F, set T to merge different fractions, if merged a meta file is required.}
}
\value{
Merged protein expression matrix
}
\description{
Use this function to merge data from different fractions
}
\examples{

dia<-dia_quant(test3,'/scratch/lb4489/project/van_cell/cety_regi/file.fp-manifest',merge=F)

}
