% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_psm.R
\name{filter_dia_psm}
\alias{filter_dia_psm}
\title{Filter diann output results}
\usage{
filter_dia_psm(peptide, fastafile, report, label)
}
\arguments{
\item{peptide}{peptide file from fragpipe}

\item{fastafile}{Fasta files for known proteins}

\item{report}{diann output}

\item{label}{Tags are used to filter known proteins, for Swiss-Prot proteins can use 'sp', It is also possible to use species names e.g. 'HUMAN', 'MOUSE', as long as it is possible to distinguish between classical and non-classical proteins.}
}
\value{
A filtered diann quantitative result, which can be used for subsequent analysis using the diann package.
}
\description{
Filter diann output results
}
\examples{

test3<-filter_dia_psm('/scratch/lb4489/project/van_cell/cety_regi/fragpipe/peptide.tsv',
                     '/scratch/lb4489/bioindex/uniprot_mouse.fasta',
                     '/scratch/lb4489/project/van_cell/cety_regi/fragpipe/diann-output/report.tsv',
                     'sp')

}
