% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_psm.R
\name{filter_psm}
\alias{filter_psm}
\title{filter Peptide}
\usage{
filter_psm(
  combined_peptide,
  combined_modified_peptide,
  fastafile,
  peptide_quant,
  datatype,
  label
)
}
\arguments{
\item{combined_peptide}{Combined_peptide file from fragpipe}

\item{combined_modified_peptide}{combined_modified_peptide file from fragpipe}

\item{fastafile}{Fasta files for known proteins}

\item{peptide_quant}{fragpipe outputs a quantitative file of peptide}

\item{datatype}{tmt or lbf}

\item{label}{Tags are used to filter known proteins, for uniprot proteins use 'sp'}
}
\value{
Quantification of peptide intensity after filtration
}
\description{
filter Peptide
}
