% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fdr.R
\name{plot_fdr}
\alias{plot_fdr}
\title{Visualisation of subgroup FDR control}
\usage{
plot_fdr(target, decoy, label)
}
\arguments{
\item{target}{target psm file, e.g. 10NIH_M_451_032523_percolator_target_psms.tsv}

\item{decoy}{decoy psm file, e.g. 10NIH_M_451_032523_percolator_decoy_psms.tsv}

\item{label}{Label for the classical proteome, 'sp' when using the uniprot proteome}
}
\value{
A list consists of three ggplot images.
}
\description{
Visualisation of subgroup FDR control
}
