% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_analysis.R
\name{coexp_gsea}
\alias{coexp_gsea}
\title{Co-expression analysis using the GSEA algorithm}
\usage{
coexp_gsea(array, name, species)
}
\arguments{
\item{array}{Protein intensity array}

\item{name}{Name of peptide analysed}

\item{species}{Only 'human' or 'mouse' is supported.}
}
\value{
A list consists of [1] a clusterProfiler object, [2] a GSEA diagram
}
\description{
Co-expression analysis using the GSEA algorithm
}
