% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_psm.R
\name{dia_quant}
\alias{dia_quant}
\title{Use this function to merge data from different fractions}
\usage{
dia_quant(quant, meta, ...)
}
\arguments{
\item{quant}{filter_dia_psm output}

\item{meta}{fp-manifest file}

\item{...}{}
}
\value{
Merged protein expression matrix
}
\description{
Use this function to merge data from different fractions
}
\seealso{
\code{\link{diann::diann_maxlfq}}
}
