% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_psm.R
\name{filter_psm}
\alias{filter_psm}
\title{filter Peptide}
\usage{
filter_psm(
  combined_peptide,
  combined_modified_peptide,
  fastafile,
  peptide_quant,
  datatype,
  label
)
}
\arguments{
\item{combined_peptide}{Combined_peptide file from fragpipe}

\item{combined_modified_peptide}{combined_modified_peptide file from fragpipe, lbf data no need}

\item{fastafile}{Fasta files for known proteins}

\item{peptide_quant}{fragpipe outputs a quantitative file of peptide, for label-free data please use the msstats.csv file}

\item{datatype}{tmt or lbf}

\item{label}{Tags are used to filter known proteins, for Swiss-Prot proteins can use 'sp', It is also possible to use species names e.g. 'HUMAN', 'MOUSE', as long as it is possible to distinguish between classical and non-classical proteins.}
}
\value{
Quantification of peptide intensity after filtration
}
\description{
filter Peptide
}
\examples{

#TMT data

test<-filter_psm('/scratch/lb4489/project/MS_data/test_fdr/tmt_313/combined_peptide.tsv',
                '/scratch/lb4489/project/MS_data/test_fdr/tmt_313/combined_modified_peptide.tsv',
                '/scratch/lb4489/bioindex/uniprot_human.fasta',
                '/scratch/lb4489/project/MS_data/test_fdr/tmt_313/tmt-report/abundance_peptide_MD.tsv',
                'tmt',
                'sp')

#label-free data

test2<-filter_psm('/scratch/lb4489/project/MS_data/test_fdr/lbf_3_14/combined_peptide.tsv',
                 '/scratch/lb4489/project/MS_data/test_fdr/lbf_3_14/combined_modified_peptide.tsv',
                 '/scratch/lb4489/bioindex/uniprot_human.fasta',
                 '/scratch/lb4489/project/MS_data/test_fdr/lbf_3_14/MSstats.csv',
                 'lbf',
                 'sp')


}
