% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_annotations.R
\name{generate_annotations}
\alias{generate_annotations}
\title{Generate annotation files for Ribo-seq analysis}
\usage{
generate_annotations(annotate_files, assemble_transcriptome)
}
\arguments{
\item{annotate_files}{Annotation file, such as gencode.v44.annotation.gff3}

\item{assemble_transcriptome}{GffCompare results}
}
\value{
Adjusted assembled transcriptome that can be merged with annotation files for Ribo-seq analysis
}
\description{
Generate annotation files for Ribo-seq analysis
}
\examples{

generate_annotations(annotate_files='gencode.v44.annotation.gff3',
assemble_transcriptome='1fpkm.annotated.gtf')


}
