% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_index.R
\name{generate_index}
\alias{generate_index}
\title{Construction of a proteogenomics index based on ribo-seq results}
\usage{
generate_index(file, source, length, label, OrganismName, OrganismIdentifier)
}
\arguments{
\item{file}{Ribo-seq analysis output file}

\item{source}{ribocode, ribotish or other}

\item{length}{Peptide length, used to filter long proteins in the index, set to 0 when there is no need to filter by length}

\item{label}{fasta head tag, e.g. "sPep", not 'sp' or 'tr' is recommended.}

\item{OrganismName}{OrganismName is the scientific name of the organism for the UniProtKB entry, defaulting to "Homo sapiens".}

\item{OrganismIdentifier}{OrganismIdentifier is the unique identifier of the source organism, assigned by NCBI, defaulting to "9696".}
}
\value{
A Biostrings object can be written as a fasta file
}
\description{
Construction of a proteogenomics index based on ribo-seq results
}
\examples{
generate_index(file='ribocode.txt',source='ribocode',length='100',label='sPep')


}
