% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fdr.R
\name{plot_fdr}
\alias{plot_fdr}
\title{Visualisation of subgroup FDR control}
\usage{
plot_fdr(target, decoy, label)
}
\arguments{
\item{target}{target psm file, e.g. 10NIH_M_451_032523_percolator_target_psms.tsv}

\item{decoy}{decoy psm file, e.g. 10NIH_M_451_032523_percolator_decoy_psms.tsv}

\item{label}{Label for the classical proteome, 'sp' when using the uniprot proteome}
}
\value{
A list consists of three ggplot images.
}
\description{
Visualisation of subgroup FDR control
}
\examples{

plot<-plot_fdr('/scratch/lb4489/project/van_cell/cety_regi/fragpipe/210323_MvO_2009_BR3_Camk2a-BULB-MIX-3_percolator_target_psms.tsv',
                '/scratch/lb4489/project/van_cell/cety_regi/fragpipe/210323_MvO_2009_BR3_Camk2a-BULB-MIX-3_percolator_decoy_psms.tsv',
                 'sp')

patchwork::wrap_plots(plot,nrow=1, guides="collect")


}
