% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fdr.R
\name{plot_peptide_fdr}
\alias{plot_peptide_fdr}
\title{Visualising the distribution of scores for specific peptides}
\usage{
plot_peptide_fdr(target, decoy, label, peptidelist)
}
\arguments{
\item{target}{target psm file, e.g. 10NIH_M_451_032523_percolator_target_psms.tsv}

\item{decoy}{decoy psm file, e.g. 10NIH_M_451_032523_percolator_decoy_psms.tsv}

\item{label}{Label for the classical proteome, 'sp' when using the uniprot proteome}

\item{peptidelist}{A sequence of peptides, e.g. peptidelist<-c('LIEFCISSMRN','MTMLADHAARQ')}
}
\value{
A list consists of two ggplot images.
}
\description{
Visualising the distribution of scores for specific peptides
}
\examples{
plot<-plot_peptide_fdr('/scratch/lb4489/project/van_cell/cety_regi/fragpipe/210323_MvO_2009_BR3_Camk2a-BULB-MIX-3_percolator_target_psms.tsv',
                       '/scratch/lb4489/project/van_cell/cety_regi/fragpipe/210323_MvO_2009_BR3_Camk2a-BULB-MIX-3_percolator_decoy_psms.tsv',
                       'sp',c('ASGGGVPTDEEQATGLER','IEREPEDNDYLWR'))

patchwork::wrap_plots(plot,nrow=2, guides="collect")



}
